+++ src/vuex/types.js
@@ -212,7 +212,7 @@ const getters = {
   fiatAssets: 'fiatAssets',
   depositableAssets: 'depositableAssets',
   coinpaymentsAssets: 'coinpaymentsAssets',
-  transferableBalancesAssets: 'transferableBalancesAssets',
+  transferableBalancesAssetsByOwner: 'transferableBalancesAssetsByOwner',
   withdrawableBalancesAssets: 'withdrawableBalancesAssets',
   statsQuoteAsset: 'statsQuoteAsset',
   ownedAssets: 'ownedAssets',